mixkey: Add a write-back cache to the replay filter.
This now uses a bloom filter along with a write-back cache for
concurrency and performance.  The bloom filter is sized such that it can
comfortably accomodate a saturated Gig-E link, but will need to be
re-adjusted for anything considerably faster.

It may be worth revisiting the flush rate, or thinking about if it is ok
to just drop packets on a false positive (p = 0.001).

Fixes #2.